icu-note-copilot/pyproject.toml
[project]
name = "icu-copilot"
version = "0.1.0"
description = "Evidence-first ICU note summarization + differential with traceability"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "numpy>=1.26",
  "pandas>=2.2",
  "rank-bm25>=0.2.2",
  "sentence-transformers>=3.0",
  "faiss-cpu>=1.8.0",
  "rapidfuzz>=3.9",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6.0"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
